requirements.txt
pillow-heif>=0.16.0
Pillow>=10.0.0
easyocr>=1.7.1
opencv-python>=4.9.0
numpy>=1.24.0
python-dateutil>=2.9.0
tqdm>=4.66.0
pytesseract>=0.3.10

